Beta commit to integrate quantum into the openstack dashboard.
This commit uses the quantum client library from Tyler Smith's
unmerged branch. The current code is checked out of a private branch
in Launchpad (to be replaced by the main quantum trunk once merged).

Installation directions are the same as the dashboard's.

If merging into an existing branch, you would need to install
the virtual environment again to load the quantum code and the
django-quantum libraries.

The local_settings.py.example file has been updated, remember to fill
in the quantum specific values in local_settings.py